refactor(web): inject member dialog analytics
